Only allow delete from library if user has DB write permission

+++ xbmc/video/windows/GUIWindowVideoNav.cpp
@@ -1046,13 +1046,21 @@ bool CGUIWindowVideoNav::OnClick(int iItem)
   if (!item->m_bIsFolder && item->IsVideoDb() && !item->Exists())
   {
     CLog::Log(LOGDEBUG, "%s called on '%s' but file doesn't exist", __FUNCTION__, item->GetPath().c_str());
-    if (!CGUIDialogVideoInfo::DeleteVideoItemFromDatabase(item, true))
-      return true;
+    if (CProfilesManager::Get().GetCurrentProfile().canWriteDatabases())
+    {
+      if (!CGUIDialogVideoInfo::DeleteVideoItemFromDatabase(item, true))
+        return true;
 
-    // update list
-    Refresh(true);
-    m_viewControl.SetSelectedItem(iItem);
-    return true;
+      // update list
+      Refresh(true);
+      m_viewControl.SetSelectedItem(iItem);
+      return true;
+    }
+    else
+    {
+      CGUIDialogOK::ShowAndGetInput(257, 0, 662, 0);
+      return true;
+    }	  
   }
   else if (StringUtils::StartsWithNoCase(item->GetPath(), "newtag://"))
   {
